Rename deterministic test modules
